aplot-package {aplot}	R Documentation

aplot: Decorate a 'ggplot' with Associated Information

Description

For many times, we are not just aligning plots as what 'cowplot' and 'patchwork' did. Users would like to align associated information that requires axes to be exactly matched in subplots, e.g. hierarchical clustering with a heatmap. This package provides utilities to aligns associated subplots to a main plot at different sides (left, right, top and bottom) with axes exactly matched.

Author(s)

Maintainer: Guangchuang Yu guangchuangyu@gmail.com (ORCID)

Other contributors:

Shuangbin Xu xshuangbin@163.com (ORCID) [contributor]

Thomas Hackl thackl@mit.edu [contributor]

See Also

Useful links:

https://github.com/YuLab-SMU/aplot

https://yulab-smu.top/aplot/

[Package aplot version 0.2.3 Index]