xlab2 {aplot}	R Documentation

display x or y axis label as an ordinary text, so that the label will not be aligned with axis label of another plot

Description

display x or y axis label as an ordinary text, so that the label will not be aligned with axis label of another plot

Usage

xlab2(label, fontsize = 11, ...)

ylab2(label, fontsize = 11, ...)

Arguments

label
axis label

fontsize
fontsize of the label

...
additional parameter passed to 'gpar'

Value

gg object with new label

Author(s)

Guangchuang Yu

[Package aplot version 0.2.3 Index]