plot_list {aplot}	R Documentation

plot a list of ggplot objects

Description

plot a list of ggplot objects using patchwork, similar to 'cowplot::plot_grid(plotlist)'

Usage

plot_list(
  ...,
  gglist = NULL,
  ncol = NULL,
  nrow = NULL,
  byrow = NULL,
  widths = NULL,
  heights = NULL,
  guides = NULL,
  labels = NULL,
  tag_levels = NULL,
  tag_size = 14,
  design = NULL,
  output = "patchwork"
)

Arguments

...
list of plots to be arranged

gglist
(optional) list of plots

ncol
number of columns

nrow
number of rows

byrow
If "FALSE" the plots will be filled in in column-major order

widths
relative widths

heights
relative heights

guides
A string specifying how guides should be treated in the layout.

labels
manual specified labels to label plots

tag_levels
format to label plots, will be disable if 'labels' is not NULL

tag_size
size of tags

design
specification of the location of areas in the layout

output
one of 'gglist' or 'patchwork'

Value

composite plot

Author(s)

Guangchuang Yu

[Package aplot version 0.2.3 Index]